requirements.txt
asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.9.2
django-rest-framework==0.1.0
djangorestframework==3.12.1
mysqlclient==2.0.1
Pillow==7.2.0
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
